Refactor: api 로직 컴포넌트에서 작성
